+++ sentry/src/main/java/io/sentry/SentryEnvelopeItemHeader.java
@@ -14,6 +14,7 @@ import org.jetbrains.annotations.Nullable;
 public final class SentryEnvelopeItemHeader implements JsonSerializable, JsonUnknown {
 
   private final @Nullable String contentType;
+  private final @Nullable Integer itemCount;
   private final @Nullable String fileName;
   private final @Nullable String platform;
   private final @NotNull SentryItemType type;
@@ -58,7 +59,8 @@ public final class SentryEnvelopeItemHeader implements JsonSerializable, JsonUnk
       final @Nullable String contentType,
       final @Nullable String fileName,
       final @Nullable String attachmentType,
-      final @Nullable String platform) {
+      final @Nullable String platform,
+      final @Nullable Integer itemCount) {
     this.type = Objects.requireNonNull(type, "type is required");
     this.contentType = contentType;
     this.length = length;
@@ -66,6 +68,7 @@ public final class SentryEnvelopeItemHeader implements JsonSerializable, JsonUnk
     this.getLength = null;
     this.attachmentType = attachmentType;
     this.platform = platform;
+    this.itemCount = itemCount;
   }
 
   SentryEnvelopeItemHeader(
@@ -74,7 +77,7 @@ public final class SentryEnvelopeItemHeader implements JsonSerializable, JsonUnk
       final @Nullable String contentType,
       final @Nullable String fileName,
       final @Nullable String attachmentType) {
-    this(type, getLength, contentType, fileName, attachmentType, null);
+    this(type, getLength, contentType, fileName, attachmentType, null, null);
   }
 
   SentryEnvelopeItemHeader(
@@ -83,7 +86,8 @@ public final class SentryEnvelopeItemHeader implements JsonSerializable, JsonUnk
       final @Nullable String contentType,
       final @Nullable String fileName,
       final @Nullable String attachmentType,
-      final @Nullable String platform) {
+      final @Nullable String platform,
+      final @Nullable Integer itemCount) {
     this.type = Objects.requireNonNull(type, "type is required");
     this.contentType = contentType;
     this.length = -1;
@@ -91,6 +95,7 @@ public final class SentryEnvelopeItemHeader implements JsonSerializable, JsonUnk
     this.getLength = getLength;
     this.attachmentType = attachmentType;
     this.platform = platform;
+    this.itemCount = itemCount;
   }
 
   SentryEnvelopeItemHeader(
@@ -119,6 +124,7 @@ public final class SentryEnvelopeItemHeader implements JsonSerializable, JsonUnk
     public static final String ATTACHMENT_TYPE = "attachment_type";
     public static final String LENGTH = "length";
     public static final String PLATFORM = "platform";
+    public static final String ITEM_COUNT = "item_count";
   }
 
   @Override
@@ -137,6 +143,9 @@ public final class SentryEnvelopeItemHeader implements JsonSerializable, JsonUnk
     }
     if (platform != null) {
       writer.name(JsonKeys.PLATFORM).value(platform);
+    }
+    if (itemCount != null) {
+      writer.name(JsonKeys.ITEM_COUNT).value(itemCount);
     }
     writer.name(JsonKeys.LENGTH).value(getLength());
     if (unknown != null) {
@@ -161,6 +170,7 @@ public final class SentryEnvelopeItemHeader implements JsonSerializable, JsonUnk
       int length = 0;
       String attachmentType = null;
       String platform = null;
+      Integer itemCount = null;
       Map<String, Object> unknown = null;
 
       while (reader.peek() == JsonToken.NAME) {
@@ -184,6 +194,9 @@ public final class SentryEnvelopeItemHeader implements JsonSerializable, JsonUnk
           case JsonKeys.PLATFORM:
             platform = reader.nextStringOrNull();
             break;
+          case JsonKeys.ITEM_COUNT:
+            itemCount = reader.nextIntegerOrNull();
+            break;
           default:
             if (unknown == null) {
               unknown = new HashMap<>();
@@ -197,7 +210,7 @@ public final class SentryEnvelopeItemHeader implements JsonSerializable, JsonUnk
       }
       SentryEnvelopeItemHeader sentryEnvelopeItemHeader =
           new SentryEnvelopeItemHeader(
-              type, length, contentType, fileName, attachmentType, platform);
+              type, length, contentType, fileName, attachmentType, platform, itemCount);
       sentryEnvelopeItemHeader.setUnknown(unknown);
       reader.endObject();
       return sentryEnvelopeItemHeader;